rename dialog moved to LogExpert.UI
